requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.27
psycopg[binary]==3.2.13
pydantic==2.11.9
email-validator==2.2.0
httpx==0.27.0
PyJWT==2.12.0
cryptography==46.0.7
argon2-cffi==23.1.0
pytest==8.3.5
